+++ pytorch_segmentation_models_trainer/custom_losses/base_loss.py
@@ -23,6 +23,7 @@
 
 import math
 from functools import partial
+from typing import List, Union
 
 import scipy.interpolate
 import numpy as np
@@ -31,6 +32,8 @@ import torch
 import torch.distributed
 from torch.nn import functional as F
 import logging
+
+from hydra.utils import instantiate
 
 from pytorch_segmentation_models_trainer.utils import math_utils
 from pytorch_segmentation_models_trainer.custom_metrics import metrics
@@ -193,11 +196,12 @@ class MultiLoss(torch.nn.Module):
 class SegLoss(Loss):
     def __init__(
         self,
-        name,
-        gt_channel_selector,
-        bce_coef=0.5,
-        dice_coef=0.5,
-        use_mixed_precision=False,
+        name: str,
+        gt_channel_selector: int,
+        bce_coef: float = 0.5,
+        dice_coef: float = 0.5,
+        tversky_focal_coef: float = 0,
+        use_mixed_precision: bool = False,
     ):
         """
         :param name:
@@ -208,9 +212,10 @@ class SegLoss(Loss):
         self.gt_channel_selector = gt_channel_selector
         self.bce_coef = bce_coef
         self.dice_coef = dice_coef
+        self.tversky_focal_coef = tversky_focal_coef
         self.use_mixed_precision = use_mixed_precision
 
-    def compute(self, pred_batch, gt_batch):
+    def compute(self, pred_batch: torch.Tensor, gt_batch: torch.Tensor):
         """
         seg and gt_polygons_image do not necessarily have the same channel count.
         gt_selector is used to select which channels of gt_polygons_image to use.
@@ -224,8 +229,6 @@ class SegLoss(Loss):
         pred_seg = pred_batch["seg"]
         gt_seg = gt_batch["gt_polygons_image"][:, self.gt_channel_selector, ...]
         weights = gt_batch["seg_loss_weights"][:, self.gt_channel_selector, ...]
-        dice = metrics.dice_loss(pred_seg, gt_seg)
-        mean_dice = torch.mean(dice)
         # if mixed precision is used, the right function is binary_cross_entropy_with_logits
         # because it can handle operations with fp16 and fp32
         cross_entropy_func = (
@@ -236,7 +239,23 @@ class SegLoss(Loss):
         mean_cross_entropy = cross_entropy_func(
             pred_seg, gt_seg, weight=weights, reduction="mean"
         )
-        return self.bce_coef * mean_cross_entropy + self.dice_coef * mean_dice
+
+        dice = metrics.dice_loss(pred_seg, gt_seg)
+        mean_dice = torch.mean(dice)
+
+        focal_tversky = (
+            0
+            if self.tversky_focal_coef == 0
+            else metrics.focal_tversky_loss(
+                pred_seg, gt_seg, gamma=0.25, alpha=0.01, beta=0.99
+            )
+        )
+        mean_focal = torch.mean(focal_tversky)
+        return (
+            self.bce_coef * mean_cross_entropy
+            + self.dice_coef * mean_dice
+            + self.tversky_focal_coef * mean_focal
+        )
 
 
 class CrossfieldAlignLoss(Loss):
@@ -450,6 +469,9 @@ def build_combined_loss(cfg):
                 gt_channel_selector=gt_channel_selector,
                 bce_coef=cfg.loss_params.seg_loss_params.bce_coef,
                 dice_coef=cfg.loss_params.seg_loss_params.dice_coef,
+                tversky_focal_coef=0
+                if "tversky_focal_coef" not in cfg.loss_params.seg_loss_params
+                else cfg.loss_params.seg_loss_params.tversky_focal_coef,
                 use_mixed_precision=True
                 if "precision" in cfg.pl_trainer and cfg.pl_trainer.precision == 16
                 else False,